endpoint: make policy calculation log per-endpoint Debug level
This log message will get emitted every time policy is calculated for an endpoint. This means that
this log will be emitted on a per-endpoint, per-policy calculation basis for an endpoint. This
will fill up logs quickly in a high-volume environment, so it's best to put it at debug level.

Signed-off by: Ian Vernon <ian@cilium.io>