CI: Stop monitor after all test assertions
[ upstream commit 625f93677695b33cd4c6e53511ddfe86cbc7d00c ]

It is possible that the cancel call we mae before checking the output
from monitor causes it to stop emitting events too soon. This moves the
cancel, thus the close, to after we definitively have passed or failed
that part of the test.

We re-enable a disabled monitor test since we haven't seen that specific
flake with this change.

Signed-off-by: Ray Bejjani <ray@covalent.io>
Signed-off-by: André Martins <andre@cilium.io>