policy: add interface for receiving updates on starting / stopping use of a selector
This new interface, `identityPopulator`, will be used for when FQDNSelectors are
created or removed from the SelectorCache. The FQDN subsystem in Cilium will
implement this interface in a future commit so that when an FQDNSelector is used
(as signaled by its addition in the SelectorCache) and removed (as signaled by
its deletion from the SelectorCache), the FQDN subsystem can be sure to check
whether the set of DNS names it is managing match that selector. This ensures
that the FQDN package only cares about FQDNSelectors which are actually being
used by endpoints on the local node, instead of all FQDNSelectors in the rules
in the policy repository.

Signed-off by: Ian Vernon <ian@cilium.io>