docs: Add note about running cilium-rm-svc-v2 when upgrading
This is needed for users who want to do the v1.5 -> v1.4 -> v1.5
upgrade.

Main reason for requiring this is that in v1.4 we do not maintain
the svc-v2 BPF maps, so they become stale. When doing the service
restoration from legacy to v2, we do not compare endpoint entries
of the maps (neither does `syncLBMapsWithK8s`), so it's better
to remove the v2 maps to force full restoration from the legacy maps.

Signed-off-by: Martynas Pumputis <m@lambda.lt>