cilium: encrypt, use ipcache to lookup IPsec destination IP
In direct routing mode without CNI chaining Cilium defaults to
allocating IPs from a IP/Mask assigned to the node. On the
encryption side we can use this to build an IPsec packet that
will be routed correctly to the node. To do this we build the
IPsec state/policy rules so that the internal IP (IP assigned
to cilium_host) for the node is used for the dip of the tunnel
header.

But, with CNI chaining we can no longer infer the destination
node from the original packets destination IP because IPs are
not assigned using a pool per node. To work around this we use
a place holder destination IP during the encryption process then
use the ipcache entry to lookup the required destination IP.

This requires routing through cilium_host for the rewrite.

Signed-off-by: John Fastabend <john.fastabend@gmail.com>