pkg/kvstore: introduced a dedicated session for locks
If a cilium agent was restarted while holding a
distributed TTL in the KVStore it would make that lock
from only being released after the default TTL of the
Cilium agent expired, 15 minutes. This commit introduces
a new session dedicated for locks which will have a TTL
of 25 seconds making the lock to be released after 25
seconds if the cilium agent is restarted.

Signed-off-by: André Martins <andre@cilium.io>